jwt-unverified-signature/requirements.txt
Flask==2.0.3
werkzeug==2.0.3
PyJWT==2.1.0

